feat: 각 배너로 text 이동
